requirements.txt
celery==4.3.0
Flask==1.0.2
flask-redis==0.4.0
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
Jinja2==2.10.1
Pillow==6.1.0
redis==3.2.1
requests==2.21.0
SQLAlchemy==1.3.5
sqlalchemy-migrate==0.12.0
Werkzeug==0.15.2
WTForms==2.2.1